SDIT-2804 Improve batch job tracing and fix Spring bean warning
Tracing: previously each step in the batch job was running in a different span so had different OperationIds in App Insights. They now all have the same OperationId, have a better OperationName (BatchManager.runBatchJob) and have a dependency that includes the `batchType`.

BeanFactoryPostProcessor: Spring was issuing a warning about creating a BeanFactoryPostProcessor as a non-static method in a Configuration class which can cause container lifecycle issues. Moved to a @Component instead and Spring is happy.
